Split up testMiscArray.py to make tests easier to read.
